Move event test fixtures to a subdirectory
