+++ tests/phpunit/event/fixtures/TestChildEvent.php
@@ -21,8 +21,8 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event;
+namespace pocketmine\event\fixtures;
 
-class TestConcreteEvent extends Event{
+class TestChildEvent extends TestParentEvent{
 
 }
