+++ tests/phpunit/event/fixtures/TestConcreteEvent.php
@@ -21,8 +21,10 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event;
+namespace pocketmine\event\fixtures;
 
-class TestConcreteExtendsAllowHandleEvent extends TestAbstractAllowHandleEvent{
+use pocketmine\event\Event;
+
+class TestConcreteEvent extends Event{
 
 }
