+++ tests/phpunit/event/fixtures/TestConcreteExtendsAbstractEvent.php
@@ -21,8 +21,8 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event;
+namespace pocketmine\event\fixtures;
 
-class TestGrandchildEvent extends TestChildEvent{
+class TestConcreteExtendsAbstractEvent extends TestAbstractEvent{
 
 }
