+++ tests/phpunit/event/fixtures/TestConcreteExtendsAllowHandleEvent.php
@@ -21,11 +21,8 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event;
+namespace pocketmine\event\fixtures;
 
-/**
- * @allowHandle
- */
-abstract class TestAbstractAllowHandleEvent extends Event{
+class TestConcreteExtendsAllowHandleEvent extends TestAbstractAllowHandleEvent{
 
 }
