+++ tests/phpunit/event/fixtures/TestConcreteExtendsConcreteEvent.php
@@ -21,8 +21,8 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event;
+namespace pocketmine\event\fixtures;
 
-class TestConcreteExtendsAbstractEvent extends TestAbstractEvent{
+class TestConcreteExtendsConcreteEvent extends TestConcreteEvent{
 
 }
